+++ src/translations/en/neo.php
@@ -27,6 +27,7 @@ return [
     'The SVG icon to use for this block type.' => 'The SVG icon to use for this block type.',
     'None set' => 'None set',
     'No icons are available to select.' => 'No icons are available to select.',
+    'Color' => 'Color',
     'Add' => 'Add',
     'Replace' => 'Replace',
     'Remove' => 'Remove',